# enable bf16 vec copy